requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
face-recognition==1.3.0
opencv-python-headless==4.7.0.72
numpy==1.24.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
